+++ aasrp/management/commands/aasrp_migrate_srp_data.py
@@ -1,5 +1,3 @@
-# coding=utf-8
-
 """
 Migrate srp data from the built-in SRP module
 """