+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
 
+## [1.2.0] - 2021-03-14
+
+### Added
+
+- Option for users to disable notifications for this module. When enabled, no
+  notifications will be sent at all. Neither in AA itself nor on Discord (if any of
+  the Discord apps is active)
+- Check if [AA Discord Notify](https://gitlab.com/ErikKalkoken/aa-discordnotify) is
+  installed, which picks up on notifications in Auth and relays them to the
+  respective user as Discord PM. This way we don't send double notifications to the
+  user if [AA-Discordbot](https://github.com/pvyParts/allianceauth-discordbot) is
+  installed as well.
+
+### Changed
+
+- Enabled paging for SRP link table on dashboard in "View All" mode
+
+
 ## [1.1.0] - 2021-03-02
 
 ### Added